Convert the Wikiizer from a streamy system to a "process it all" mechanism. Wiki pages "shouldn't" be that large, so this should be quite fine (and even so, it's memory footprint should scale linearly with the input).
This enabled some decent code reduction (e.g. toss BufferedInput).

I also noticed that the Wikiizer still refered to the global closers,
so I made it possible to override those, too.

Lastly, I added a small feature: a processor can return an object for
insertion into the output text. That object will have __str__ called
on it after the closers, and just before being returned. This enables
a processor to defer the evaluation of the result text (e.g. so that
the processor can collect h2 tags from the page to construct a TOC at
the top of the page).

* lib/subwiki/wikiize.py:
  (Wikiizer.__init__): pass closers. simplify the logic for selecting
    the parameter-based or global processors (and closers). forget the
    buffered input and pending text -- just evaluate the input and
    stash it away for (later) streamy reading.
  (Wikiizer.read): consume values from self._result. most of the old
    content was moved to the new Wikiizer.process method.
  (Wikiizer.process): new method, derived from the old .read method.
    added the capability to directly pass processors or closers, and
    to default to the instance values for these. change the look to
    just split the provided text instead of reading lines. simplify
    the logic since we know that we'll always run closers, rather than
    just when we hit the end of the input stream. apply str() to each
    element to support deferred evaluation.
  (_BUFFERSIZE): removed
  (class BufferedInput): removed
  (...): removed import of cStringIO; no longer necessary