The old code assumed the leadings parts of a ref had to be sequences if trailing parts were to be sequences, too. However, this failed for a sequence which was stored in an object. nonseq.seq would throw an error.
The new code will now look at the reference and see if some part of it
refers to an active [for] reference. If so, then it returns the
appropriate construction for later reference to the iterative value.

The problem identification and an initial patch were provided by Sander
Striker <striker@apache.org>, then evolved by gstein.

* ezt.py:
  (_prepare_ref): look for a [for] reference with any part of the
    reference, rather than assuming all 0..n parts are sequences.

* tests/test.py: a simple test to check this case
* tests/test.ezt: a template for test.py

* tests/timing.py: short dev script to try different algorithms for
    determining if a reference contains an active [for] identifier.

* properties on '.': set svn:ignore to ignore *.pyc