Patch from Leonard Richardson <leonardr@linux.ucla.edu>. He writes:
This patch gets rid of the displayed "mode" of a page being viewed, as
opposed to edited or history-viewed. My thinking is that someone who's
just browsing a wiki is not in "mode mode", as it were. A user only
becomes interested in which mode they're in when they deviate from the
main purpose of the wiki, which is to display information.

Precedent: Wikipedia has headings like "Xenobiology" and "Editing
Xenobiology", not "Viewing Xenobiology" and "Editing Xenobiology".

Changes:

* header.ezt: Suppress the mode descriptor in the title and heading when
  no mode name is given.

* view.ezt: Since to "view" is the default mode, pass in an empty string
  for the mode name.