ne_defs.h: Add NE_VAR macro
This adds a macro to decorate public variables that is defined as 'extern' when
building on non-Visual Studio compilers or when libneon is built statically.

However when libneon is built or used as a shared (DLL) build, the macro will
be defined to be 'extern __declspec(dllexport)' and
'extern __declspec(dllimport)' so that the variables are correctly exported and
imported from the DLL during linking.