neon.mak: Support building against libxml2 too
Allow the build to support building against libxml2 too, if one explicitly
requests for it and not request Expat.  This means that if both are requested,
Expat is preferred over libxml2.

Drop the parts where we try to integrate the build with the Expat sources since
it seems that the it is no longer bundled and the Visual Studio build files and
layout has changed quite a bit as it switched to CMake.

Update entries in the NMake Makefile to support both types of XML support.